Adjusted the clearConsole function
Using u8s for the sizes causes the compiled code to truncate the upper 24 bits of the values of the sizes every time that the loop runs, which uses more memory than necessary.

+++ include/global.h
@@ -4,7 +4,7 @@
 
 extern "C"
 {
-	void clearConsole(u8 lines);
+	void clearConsole(u32 lines);
 	bool indexOf(u8 array[], u32 size, u8 value);
 	bool checkForButtonInput(u32 buttonCombo);
 	void setConsoleColor(u32 rgba);
